# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synops Praxis LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /studio/generate-script:
    post:
      operationId: generateScript
      tags: [studio]
      summary: Generate script beats from source document text (Claude)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScriptInput"
      responses:
        "200":
          description: Generated script draft with beats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptDraft"
...
  /studio/scripts/{draftId}:
    get:
      operationId: getScriptDraft
      tags: [studio]
      summary: Get a script draft
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Script draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptDraft"
    patch:
      operationId: updateScriptDraft
      tags: [studio]
      summary: Update/edit a script draft before publishing
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptDraftUpdate"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptDraft"
  /studio/scripts/{draftId}/publish:
    post:
      operationId: publishScriptDraft
      tags: [studio]
      summary: Convert approved script draft into a module with beats
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishDraftInput"
      responses:
        "201":
          description: Module created from draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"

  # ── Socratic Learning Sessions ─────────────────────────────────────────────
...
components:
  schemas:
...
    ModuleDetail:
      type: object
      required: [id, courseId, title, status, order, beats, createdAt]
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        order:
          type: integer
        beats:
          type: array
          items:
            $ref: "#/components/schemas/Beat"
        estimatedMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ── Beats ──────────────────────────────────────────────────────────────
    Beat:
      type: object
      required: [id, moduleId, type, order, title, narration]
      properties:
        id:
          type: string
        moduleId:
          type: string
        type:
          type: string
          enum: [title_card, points, scenario, compare, diagram, close]
        order:
          type: integer
        title:
          type: string
        narration:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        scenario:
          type: ["string", "null"]
        visualData:
          description: Arbitrary JSON for template rendering
        audioUrl:
          type: ["string", "null"]
        audioStatus:
          type: string
          enum: [none, pending, ready, error]
...
    BeatUpdate:
      type: object
      properties:
        title:
          type: string
        narration:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        scenario:
          type: string
        order:
          type: integer

    # ── Studio ─────────────────────────────────────────────────────────────
    GenerateScriptInput:
      type: object
      required: [sourceText, title]
      properties:
        title:
          type: string
        sourceText:
          type: string
        language:
          type: string
          default: "en-ZA"
        maxBeats:
          type: integer
          default: 6
    ScriptDraft:
      type: object
      required: [id, title, status, beats, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        sourceText:
          type: string
        status:
          type: string
          enum: [generating, ready, published]
        beats:
          type: array
          items:
            $ref: "#/components/schemas/Beat"
        tenantId:
          type: string
        createdAt:
          type: string
    ScriptDraftUpdate:
      type: object
      properties:
        title:
          type: string
        beats:
          type: array
          items:
            $ref: "#/components/schemas/BeatUpdate"
    PublishDraftInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: string
        moduleTitle:
          type: string

    # ── Sessions ───────────────────────────────────────────────────────────
...