# lib/api-zod/src/generated/types/publishDraftInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Synops Praxis LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface PublishDraftInput {
  courseId: string;
  moduleTitle?: string;
}